# Refactoring towards GPU support