--- a/datashader/compiler.py
+++ b/datashader/compiler.py
@@ -51,7 +51,7 @@ def compile_components(agg, schema, glyph):
     reds = list(traverse_aggregation(agg))
 
     # List of base reductions (actually computed)
-    bases = list(unique(concat(r._bases for r in reds)))
+    bases = list(unique(concat(r._build_bases() for r in reds)))
     dshapes = [b.out_dshape(schema) for b in bases]
     # List of tuples of (append, base, input columns, temps)
     calls = [_get_call_tuples(b, d, schema) for (b, d) in zip(bases, dshapes)]
@@ -80,7 +80,7 @@ def traverse_aggregation(agg):
 
 
 def _get_call_tuples(base, dshape, schema):
-    return base._build_append(dshape, schema), (base,), base.inputs, base._temps
+    return base._build_append(dshape, schema), (base,), base.inputs, base._build_temps()
 
 
 def make_create(bases, dshapes):
@@ -151,7 +151,12 @@ def make_finalize(bases, agg, schema):
         calls = []
         for key, val in zip(agg.keys, agg.values):
             f = make_finalize(bases, val, schema)
-            inds = [arg_lk[b] for b in getattr(val, '_bases', bases)]
+            try:
+                # Override bases if possible
+                bases = val._build_bases()
+            except AttributeError:
+                pass
+            inds = [arg_lk[b] for b in bases]
             calls.append((key, f, inds))
 
         def finalize(bases, **kwargs):