--- a/datashader/data_libraries/__init__.py
+++ b/datashader/data_libraries/__init__.py
@@ -1,0 +1,7 @@
+from . import pandas, xarray  # noqa (API import)
+
+try:
+    import dask as _dask  # noqa (Test dask installed)
+    from . import dask    # noqa (API import)
+except ImportError:
+    pass