--- a/datashader/reductions.py
+++ b/datashader/reductions.py
@@ -6,7 +6,7 @@ from datashape import coretypes as ct
 from toolz import concat, unique
 import xarray as xr
 
-from .utils import Expr, ngjit
+from .utils import Expr, ngjit, isrealfloat
 
 
 class Preprocess(Expr):
@@ -60,8 +60,13 @@ class Reduction(Expr):
     def _build_create(self, dshape):
         return self._create
 
-    def _build_append(self, dshape):
-        return self._append
+    def _build_append(self, dshape, schema):
+        if self.column is None:
+            return self._append_no_field
+        elif isrealfloat(schema[self.column]):
+            return self._append_float_field
+        else:
+            return self._append_int_field
 
     def _build_combine(self, dshape):
         return self._combine
@@ -82,9 +87,6 @@ class OptionalFieldReduction(Reduction):
     def validate(self, in_dshape):
         pass
 
-    def _build_append(self, dshape):
-        return self._append if self.column is None else self._append_non_na
-
     @staticmethod
     def _finalize(bases, **kwargs):
         return xr.DataArray(bases[0], **kwargs)
@@ -103,13 +105,18 @@ class count(OptionalFieldReduction):
 
     @staticmethod
     @ngjit
-    def _append(x, y, agg):
+    def _append_no_field(x, y, agg):
         agg[y, x] += 1
 
     @staticmethod
     @ngjit
-    def _append_non_na(x, y, agg, field):
-        if not np.isnan(field):
+    def _append_int_field(x, y, agg, field):
+        agg[y, x] += 1
+
+    @staticmethod
+    @ngjit
+    def _append_float_field(x, y, agg, field):
+        if not isnan(field):
             agg[y, x] += 1
 
     @staticmethod
@@ -133,13 +140,17 @@ class any(OptionalFieldReduction):
 
     @staticmethod
     @ngjit
-    def _append(x, y, agg):
+    def _append_no_field(x, y, agg):
         agg[y, x] = True
 
     @staticmethod
     @ngjit
-    def _append_non_na(x, y, agg, field):
-        if not np.isnan(field):
+    def _append_int_field(x, y, agg, field):
+        agg[y, x] = True
+
+    @ngjit
+    def _append_float_field(x, y, agg, field):
+        if not isnan(field):
             agg[y, x] = True
 
     @staticmethod
@@ -175,12 +186,14 @@ class sum(FloatingReduction):
     """
     @staticmethod
     @ngjit
-    def _append(x, y, agg, field):
-        if not np.isnan(field):
-            if np.isnan(agg[y, x]):
-                agg[y, x] = field
-            else:
-                agg[y, x] += field
+    def _append_int_field(x, y, agg, field):
+        agg[y, x] += field
+
+    @staticmethod
+    @ngjit
+    def _append_float_field(x, y, agg, field):
+        if not isnan(field):
+            agg[y, x] += field
 
     @staticmethod
     def _combine(aggs):
@@ -208,16 +221,24 @@ class m2(FloatingReduction):
 
     @staticmethod
     @ngjit
-    def _append(x, y, m2, field, sum, count):
+    def _append_float_field(x, y, m2, field, sum, count):
         # sum & count are the results of sum[y, x], count[y, x] before being
         # updated by field
-        if not np.isnan(field):
-            if count == 0:
-                m2[y, x] = 0
-            else:
+        if not isnan(field):
+            if count > 0:
                 u1 = np.float64(sum) / count
                 u = np.float64(sum + field) / (count + 1)
                 m2[y, x] += (field - u1) * (field - u)
+
+    @staticmethod
+    @ngjit
+    def _append_int_field(x, y, m2, field, sum, count):
+        # sum & count are the results of sum[y, x], count[y, x] before being
+        # updated by field
+        if count > 0:
+            u1 = np.float64(sum) / count
+            u = np.float64(sum + field) / (count + 1)
+            m2[y, x] += (field - u1) * (field - u)
 
     @staticmethod
     def _combine(Ms, sums, ns):
@@ -237,11 +258,12 @@ class min(FloatingReduction):
     """
     @staticmethod
     @ngjit
-    def _append(x, y, agg, field):
-        if np.isnan(agg[y, x]):
+    def _append_float_field(x, y, agg, field):
+        if isnan(agg[y, x]):
             agg[y, x] = field
         elif agg[y, x] > field:
             agg[y, x] = field
+    _append_int_field = _append_float_field
 
     @staticmethod
     def _combine(aggs):
@@ -259,11 +281,12 @@ class max(FloatingReduction):
     """
     @staticmethod
     @ngjit
-    def _append(x, y, agg, field):
-        if np.isnan(agg[y, x]):
+    def _append_float_field(x, y, agg, field):
+        if isnan(agg[y, x]):
             agg[y, x] = field
         elif agg[y, x] < field:
             agg[y, x] = field
+    _append_int_field = _append_float_field
 
     @staticmethod
     def _combine(aggs):
@@ -298,8 +321,9 @@ class count_cat(Reduction):
 
     @staticmethod
     @ngjit
-    def _append(x, y, agg, field):
+    def _append_int_field(x, y, agg, field):
         agg[y, x, field] += 1
+    _append_float_field = _append_int_field
 
     @staticmethod
     def _combine(aggs):