… datashader/tests/benchmarks/test_draw_line.py
… datashader/tests/benchmarks/test_extend_line.py
… datashader/tests/test_glyphs.py
… datashader/tests/test_pandas.py
… datashader/tests/test_transfer_functions.py
… datashader/transfer_functions.py
… datashader/utils.py