@jonmmease 2019-10-10
I think I've figured out a way to support `line`/`area` with `axis=0` with cuda. So I'm going to do a bit more `line`/`area` refactorting on this PR.