@jonmmease 2019-10-11
Ok, I did a bit more refactoring of lines in d7f3650 and area in b5dbf1d.

The goal was to make it possible to support parallelization with `axis=0`.  The reason that this was not as straightforward as `axis=1` is because the prior logic to calculate breaks due to nans and out-of-bounds segments was sequential in nature.

I also managed to eliminate a fair amount of code duplication by having the `draw_segment` and `draw_trapezoid` functions input values in data, rather than pixel, coordinates. These functions now perform their own skipping/clipping logic, so this logic is no longer duplicated in every line line/area rendering function.