@jonmmease 2019-10-11
Thanks, I've done several benchmarks and I'm seeing equivalent performance (with +-5%). e.g., for 10 million points with `count` aggregation

 - `points`: From 65.2 ms -> 64.3 ms (after)
 - `line`: From 256 ms -> 268 ms (after)
 - `area`: 1800 ms -> 1750 ms (after)